Enhance activation functions and streamline project structure
Updated the `IActivationFunction` interface for better documentation. Added `Linear` and `Sigmoid` classes implementing this interface for identity and Sigmoid activation functions, respectively. Modified the `Layer` class to accept an activation function in its constructor and added a `Compute` method. Updated the `PerceptronNeuron` class to utilize the `IActivationFunction` for output calculations. Removed an empty folder declaration in the project file to streamline the structure.